refactor(#602): extract SettingsView sections into SettingsViewSections.swift
Moves all section vars and helpers into an extension on SettingsView
in a companion file, matching the PopoverMainViewSubviews pattern.
SettingsView.swift retains only state, body, and lifecycle.

Closes #602
