+++ Sources/RunnerBar/SettingsView.swift
@@ -23,7 +23,7 @@ import SwiftUI
 // is major major major.
 
 // MARK: - URI Constants
-private enum SettingsURIs {
+enum SettingsURIs {
     static let privacyPolicy  = "https://dev.eon.st/runnerbar/privacy"
     static let termsOfService = "https://dev.eon.st/runnerbar/terms"
 }
@@ -31,9 +31,7 @@ private enum SettingsURIs {
 // swiftlint:disable:next type_body_length
 struct SettingsView: View {
     let onBack: () -> Void
-    /// Called when the user taps a runner row; navigates to RunnerDetailView.
     let onSelectRunner: (RunnerModel) -> Void
-    /// #499: Called when the user taps a scope row; navigates to ScopeDetailView.
     let onSelectScope: (ScopeEntry) -> Void
     @ObservedObject var store: RunnerStoreObservable
     @ObservedObject private var settings = SettingsStore.shared
@@ -42,17 +40,14 @@ struct SettingsView: View {
     @ObservedObject private var localRunnerStore = LocalRunnerStore.shared
     @ObservedObject private var scopeStore = ScopeStore.shared
     @State private var launchAtLogin = LoginItem.isEnabled
-    // isOAuthAuthenticated: true only when a token is stored in Keychain via native OAuth.
-    // isCLIAuthenticated: true when gh CLI provides a token but no Keychain OAuth token exists.
-    // These two are mutually exclusive. Sign in button is shown whenever isOAuthAuthenticated == false.
-    @State private var isOAuthAuthenticated = (Keychain.token != nil)
-    @State private var isCLIAuthenticated = (Keychain.token == nil && githubToken() != nil)
-    @State private var isSigningIn = false
-    @State private var hasLoadedOnce = false
-    @State private var runnerPendingRemoval: RunnerModel?
-    @State private var showAddRunnerSheet = false
-    @State private var showAddScopeSheet = false
-    @State private var removeErrorMessage: String?
+    @State var isOAuthAuthenticated = (Keychain.token != nil)
+    @State var isCLIAuthenticated = (Keychain.token == nil && githubToken() != nil)
+    @State var isSigningIn = false
+    @State var hasLoadedOnce = false
+    @State var runnerPendingRemoval: RunnerModel?
+    @State var showAddRunnerSheet = false
+    @State var showAddScopeSheet = false
+    @State var removeErrorMessage: String?
 
     private var appVersion: String {
         Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? "—"
@@ -60,7 +55,7 @@ struct SettingsView: View {
     private var appBuild: String {
         Bundle.main.infoDictionary?["CFBundleVersion"] as? String ?? "—"
     }
-    private var removalAlertTitle: String {
+    var removalAlertTitle: String {
         "Remove runner \"\(runnerPendingRemoval?.runnerName ?? "this runner\"")"
     }
 
@@ -68,12 +63,6 @@ struct SettingsView: View {
         VStack(alignment: .leading, spacing: 0) {
             headerBar
             Divider()
-            // maxHeight: .infinity — fills all space the panel gives us.
-            // AppDelegate caps the panel at 85% visibleFrame. That IS the limit.
-            // ❌ NEVER move headerBar inside this ScrollView.
-            // ❌ NEVER replace .infinity with a fixed number.
-            // If you are an agent or human, DO NOT REMOVE THIS COMMENT, YOU ARE NOT ALLOWED
-            // UNDER ANY CIRCUMSTANCE.
             ScrollView(.vertical, showsIndicators: true) {
                 sectionsStack
             }
@@ -88,10 +77,11 @@ struct SettingsView: View {
         .modifier(removalAlertModifier)
     }
 
-    private func addRunnerSheet() -> some View {
+    func addRunnerSheet() -> some View {
         AddRunnerSheet(isPresented: $showAddRunnerSheet) { localRunnerStore.refresh() }
     }
-    private var removalAlertModifier: RemovalAlertModifier {
+
+    var removalAlertModifier: RemovalAlertModifier {
         RemovalAlertModifier(
             title: removalAlertTitle,
             isPresented: Binding(
@@ -104,28 +94,9 @@ struct SettingsView: View {
         )
     }
 
-    private var sectionsStack: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            localRunnersSection
-            Divider()
-            remoteScopesSection
-            Divider()
-            notificationsSection
-            Divider()
-            generalSection
-            Divider()
-            accountSection
-            Divider()
-            aboutSection
-        }
-        .padding(.bottom, 16)
-    }
-
-    private func onAppearAction() {
+    func onAppearAction() {
         isOAuthAuthenticated = (Keychain.token != nil)
         isCLIAuthenticated = (Keychain.token == nil && githubToken() != nil)
-        // Register onCompletion ONCE here — do NOT re-assign in signInWithGitHub().
-        // Re-assigning mid-flow would race with an in-flight callback.
         OAuthService.shared.onCompletion = { success in
             isOAuthAuthenticated = success
             isCLIAuthenticated = !success && githubToken() != nil
@@ -135,122 +106,31 @@ struct SettingsView: View {
         localRunnerStore.refresh()
     }
 
-    // MARK: - Header
-    private var headerBar: some View {
-        HStack {
-            Button(action: onBack, label: {
-                HStack(spacing: 4) {
-                    Image(systemName: "chevron.left")
-                        .font(.system(size: 12, weight: .medium))
-                    Text("Settings").font(.headline)
-                }
-                .foregroundColor(.primary)
-            })
-            .buttonStyle(.plain)
-            Spacer()
-        }
-        .padding(.horizontal, RBSpacing.md).padding(.top, 12).padding(.bottom, 8)
+    func applyLaunchAtLogin(_ enabled: Bool) { LoginItem.setEnabled(enabled) }
+
+    func signInWithGitHub() {
+        isSigningIn = true
+        OAuthService.shared.signIn()
     }
 
-    // MARK: - Local Runners
-    private var localRunnersSectionHeader: some View {
-        HStack {
-            Text("Active local runners")
-                .font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-            Spacer()
-            Button(action: { showAddRunnerSheet = true }, label: {
-                Image(systemName: "plus").font(.caption).foregroundColor(Color.rbTextSecondary)
-            })
-            .buttonStyle(.plain).help("Add a new runner").padding(.trailing, 4)
-            if localRunnerStore.isScanning {
-                ProgressView().scaleEffect(0.6).frame(width: 14, height: 14)
-            } else {
-                Button(action: { removeErrorMessage = nil; localRunnerStore.refresh() }, label: {
-                    Image(systemName: "arrow.clockwise").font(.caption).foregroundColor(Color.rbTextSecondary)
-                })
-                .buttonStyle(.plain).help("Refresh local runner list")
-            }
-        }
-        .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+    func signOutOfGitHub() {
+        OAuthService.shared.signOut()
     }
 
-    // #512: Description below Local Runners header (mirrors Scopes section)
-    private var localRunnersSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            localRunnersSectionHeader
-            Text("Self-hosted runners installed on this machine, discovered via LaunchAgent plists.")
-                .font(.caption).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
-            if let errMsg = removeErrorMessage {
-                Text(errMsg).font(.caption).foregroundColor(Color.rbDanger)
-                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
-                    .background(Color.rbDanger.opacity(0.07))
-            }
-            if localRunnerStore.runners.isEmpty && !localRunnerStore.isScanning && hasLoadedOnce {
-                Text("No local runners found").font(.caption).foregroundColor(Color.rbTextSecondary)
-                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
-            } else {
-                ForEach(localRunnerStore.runners) { runner in localRunnerRow(runner) }
+    func performRemoval() {
+        guard let runner = runnerPendingRemoval else { return }
+        runnerPendingRemoval = nil
+        LocalRunnerStore.shared.optimisticallyRemove(runner.runnerName)
+        DispatchQueue.global(qos: .userInitiated).async {
+            let ok = RunnerLifecycleService.shared.remove(runner: runner)
+            DispatchQueue.main.async {
+                if !ok { removeErrorMessage = "Failed to remove \"\(runner.runnerName)\". Check logs." }
+                LocalRunnerStore.shared.refresh()
             }
         }
     }
 
-    private func localRunnerRow(_ runner: RunnerModel) -> some View {
-        // swiftlint:disable:next multiple_closures_with_trailing_closure
-        Button(action: { onSelectRunner(runner) }) {
-            localRunnerRowContent(runner)
-                .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .padding(.horizontal, RBSpacing.md).padding(.vertical, 5)
-        .background(
-            RoundedRectangle(cornerRadius: RBRadius.small)
-                .fill(Color.rbSurfaceElevated)
-                .overlay(RoundedRectangle(cornerRadius: RBRadius.small)
-                    .strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5))
-                .padding(.horizontal, RBSpacing.xs)
-        )
-    }
-
-    // #507: chevron is now always the last item before the remove button (far right).
-    private func localRunnerRowContent(_ runner: RunnerModel) -> some View {
-        let hasWarning = runner.lifecycleWarning != nil
-        let displayStatus = runner.displayStatus
-        return HStack(spacing: 6) {
-            Circle().fill(localRunnerDotColor(for: runner)).frame(width: 8, height: 8)
-            VStack(alignment: .leading, spacing: 1) {
-                Text(runner.runnerName).font(.system(size: 13)).lineLimit(1)
-                if let url = runner.gitHubUrl {
-                    Text(url).font(.caption2).foregroundColor(Color.rbTextSecondary).lineLimit(1)
-                }
-            }
-            Spacer()
-            Text(displayStatus)
-                .font(.caption)
-                .foregroundColor(hasWarning ? Color.rbWarning : Color.rbTextSecondary)
-                .lineLimit(1)
-                .fixedSize()
-            if runner.isRunning {
-                Button(action: { performStop(runner: runner) },
-                       label: { Text("Stop").font(.caption2) })
-                .buttonStyle(.bordered).help("Stop runner service")
-            } else {
-                Button(action: { performResume(runner: runner) },
-                       label: { Text("Resume").font(.caption2) })
-                .buttonStyle(.bordered).help("Start runner service")
-            }
-            Image(systemName: "chevron.right")
-                .font(.caption2)
-                .foregroundColor(Color.rbTextTertiary)
-            Button(action: { runnerPendingRemoval = runner },
-                   label: { Image(systemName: "minus.circle").font(.caption2).foregroundColor(Color.rbDanger) })
-            .buttonStyle(.plain).help("Remove runner")
-        }
-    }
-
-    // MARK: - Resume / Stop actions
-
-    private func performResume(runner: RunnerModel) {
+    func performResume(runner: RunnerModel) {
         log("SettingsView > performResume called runner=\(runner.runnerName)")
         LocalRunnerStore.shared.optimisticallySetRunning(runner.runnerName, isRunning: true)
         DispatchQueue.global(qos: .userInitiated).async {
@@ -272,7 +152,7 @@ struct SettingsView: View {
         }
     }
 
-    private func performStop(runner: RunnerModel) {
+    func performStop(runner: RunnerModel) {
         log("SettingsView > performStop called runner=\(runner.runnerName)")
         LocalRunnerStore.shared.optimisticallySetRunning(runner.runnerName, isRunning: false)
         DispatchQueue.global(qos: .userInitiated).async {
@@ -288,306 +168,6 @@ struct SettingsView: View {
                         .first(where: { !$0.trimmingCharacters(in: .whitespaces).isEmpty }) ?? msg
                     LocalRunnerStore.shared.setLifecycleWarning(runner.runnerName, warning: "⚠ \(short)")
                 }
-                LocalRunnerStore.shared.refresh()
-            }
-        }
-    }
-
-    private func localRunnerDotColor(for runner: RunnerModel) -> Color {
-        switch runner.statusColor {
-        case .running: return Color.rbSuccess
-        case .busy:    return Color.rbWarning
-        case .idle:    return Color.rbTextTertiary
-        case .offline: return Color.rbDanger
-        }
-    }
-
-    // MARK: - Remote runner scopes
-
-    private var remoteScopesSectionHeader: some View {
-        HStack {
-            Text("Remote runner scopes")
-                .font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-            Spacer()
-            // swiftlint:disable:next multiple_closures_with_trailing_closure
-            Button(action: { showAddScopeSheet = true }) {
-                Image(systemName: "plus").font(.caption).foregroundColor(Color.rbTextSecondary)
-            }
-            .buttonStyle(.plain).help("Add a remote scope")
-        }
-        .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 2)
-    }
-
-    private var remoteScopesSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            remoteScopesSectionHeader
-            Text("GitHub repos or orgs whose runners are fetched via the API.")
-                .font(.caption).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
-            if scopeStore.entries.isEmpty {
-                Text("No remote scopes added")
-                    .font(.caption).foregroundColor(Color.rbTextSecondary)
-                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
-            } else {
-                ForEach(scopeStore.entries) { entry in
-                    scopeRow(entry)
-                }
-            }
-        }
-    }
-
-    private func scopeRow(_ entry: ScopeEntry) -> some View {
-        let isRepo = entry.scope.contains("/")
-        let displayName = ScopeSettingsStore.displayName(for: entry.scope)
-        // swiftlint:disable:next multiple_closures_with_trailing_closure
-        return Button(action: { onSelectScope(entry) }) {
-            HStack(spacing: 8) {
-                Text(isRepo ? "Repo" : "Org")
-                    .font(.caption2)
-                    .foregroundColor(Color.rbTextSecondary)
-                    .padding(.horizontal, 5)
-                    .padding(.vertical, 2)
-                    .background(Capsule().fill(Color.rbSurfaceElevated))
-                    .overlay(Capsule().strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5))
-
-                VStack(alignment: .leading, spacing: 1) {
-                    Text(displayName)
-                        .font(.system(size: 12))
-                        .lineLimit(1)
-                        .truncationMode(.middle)
-                    if ScopeSettingsStore.alias(for: entry.scope) != nil {
-                        Text(entry.scope)
-                            .font(.caption2)
-                            .foregroundColor(Color.rbTextTertiary)
-                            .lineLimit(1).truncationMode(.middle)
-                    }
-                }
-
-                Spacer()
-
-                Text(entry.isEnabled ? "Active" : "Paused")
-                    .font(.caption2)
-                    .foregroundColor(entry.isEnabled ? Color.rbSuccess : Color.rbTextTertiary)
-
-                Toggle("", isOn: Binding(
-                    get: { entry.isEnabled },
-                    set: { ScopeStore.shared.setEnabled(entry.id, $0); RunnerStore.shared.start() }
-                ))
-                .toggleStyle(.switch)
-                .tint(Color.rbSuccess)
-                .labelsHidden()
-                .help(entry.isEnabled ? "Pause monitoring" : "Resume monitoring")
-                .scaleEffect(0.8, anchor: .trailing)
-                .buttonStyle(.borderless)
-
-                Image(systemName: "chevron.right")
-                    .font(.caption2)
-                    .foregroundColor(Color.rbTextTertiary)
-
-                // swiftlint:disable:next multiple_closures_with_trailing_closure
-                Button(action: {
-                    ScopeSettingsStore.cleanUp(scope: entry.scope)
-                    ScopeStore.shared.remove(id: entry.id)
-                    RunnerStore.shared.start()
-                }) {
-                    Image(systemName: "minus.circle")
-                        .font(.caption2)
-                        .foregroundColor(Color.rbDanger)
-                }
-                .buttonStyle(.borderless)
-                .help("Remove scope")
-            }
-            .contentShape(Rectangle())
-        }
-        .buttonStyle(.plain)
-        .padding(.horizontal, RBSpacing.md)
-        .padding(.vertical, 5)
-        .background(
-            RoundedRectangle(cornerRadius: RBRadius.small)
-                .fill(entry.isEnabled ? Color.rbSurfaceElevated : Color.rbSurface)
-                .overlay(
-                    RoundedRectangle(cornerRadius: RBRadius.small)
-                        .strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5)
-                )
-                .padding(.horizontal, RBSpacing.xs)
-        )
-        .opacity(entry.isEnabled ? 1.0 : 0.5)
-    }
-
-    // MARK: - Notifications
-    private var notificationsSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("Notifications").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
-            HStack {
-                Text("Notify on success").font(.system(size: 12)); Spacer()
-                Toggle("", isOn: $notifications.notifyOnSuccess)
-                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
-            Divider().padding(.leading, RBSpacing.md)
-            HStack {
-                Text("Notify on failure").font(.system(size: 12)); Spacer()
-                Toggle("", isOn: $notifications.notifyOnFailure)
-                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
-        }
-    }
-
-    // MARK: - General
-    private var generalSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("General").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
-            HStack {
-                Text("Launch at login").font(.system(size: 12)); Spacer()
-                Toggle("", isOn: $launchAtLogin)
-                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
-                    .onChange(of: launchAtLogin, perform: applyLaunchAtLogin)
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
-            Divider().padding(.leading, RBSpacing.md)
-            HStack {
-                Text("Polling interval").font(.system(size: 12)); Spacer()
-                Text("\(settings.pollingInterval)s").font(.system(size: 12)).foregroundColor(Color.rbTextSecondary)
-                    .frame(minWidth: 36, alignment: .trailing)
-                Stepper("", value: $settings.pollingInterval, in: 10...300).labelsHidden()
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.top, 6).padding(.bottom, 2)
-            Text("How often RunnerBar checks GitHub for runner and workflow status. Lower values use more API quota.")
-                .font(.caption).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
-        }
-    }
-
-    // MARK: - Account
-    //
-    // Four visible states:
-    //
-    //  1. Signing in (in-flight):  spinner + "Waiting for browser…"
-    //
-    //  2. OAuth (Keychain token):  ● green  "Authenticated"
-    //                              caption: "via OAuth"
-    //                              [Sign out] bordered danger button
-    //
-    //  3. gh env token (CLI):      ● green  "Authenticated"
-    //                              caption: "via gh env token"
-    //                              [Sign in with GitHub] bordered button (optional upgrade)
-    //
-    //  4. No token at all:         [Sign in with GitHub] bordered button only
-    //
-    // Both state 2 and 3 are fully authenticated with equal API access.
-    // The caption communicates the method, not the capability.
-    //
-    // Button styling:
-    //   "Sign in with GitHub" — .buttonStyle(.bordered), .font(.caption2)
-    //   "Sign out"            — .buttonStyle(.bordered), .tint(.rbDanger), .font(.caption2)
-    //
-    // ❌ NEVER revert buttons to .buttonStyle(.plain) without a visual affordance.
-    private var accountSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("Account").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
-            HStack(alignment: .center) {
-                Text("GitHub").font(.system(size: 12))
-                Spacer()
-                if isSigningIn {
-                    HStack(spacing: 6) {
-                        ProgressView().scaleEffect(0.7)
-                        Text("Waiting for browser…").font(.caption).foregroundColor(Color.rbTextSecondary)
-                    }
-                } else if isOAuthAuthenticated {
-                    // State 2: OAuth token in Keychain.
-                    HStack(spacing: 10) {
-                        HStack(spacing: 4) {
-                            Circle().fill(Color.rbSuccess).frame(width: 7, height: 7)
-                            VStack(alignment: .leading, spacing: 1) {
-                                Text("Authenticated")
-                                    .font(.caption)
-                                    .foregroundColor(Color.rbTextSecondary)
-                                Text("via OAuth")
-                                    .font(.caption2)
-                                    .foregroundColor(Color.rbTextTertiary)
-                            }
-                        }
-                        Button(action: signOutOfGitHub) {
-                            Text("Sign out").font(.caption2)
-                        }
-                        .buttonStyle(.bordered)
-                        .tint(Color.rbDanger)
-                        .help("Remove OAuth token from Keychain. gh env token used as fallback if available.")
-                    }
-                } else if isCLIAuthenticated {
-                    // State 3: gh env token present, no Keychain OAuth token.
-                    HStack(spacing: 10) {
-                        HStack(spacing: 4) {
-                            Circle().fill(Color.rbSuccess).frame(width: 7, height: 7)
-                            VStack(alignment: .leading, spacing: 1) {
-                                Text("Authenticated")
-                                    .font(.caption)
-                                    .foregroundColor(Color.rbTextSecondary)
-                                Text("via gh env token")
-                                    .font(.caption2)
-                                    .foregroundColor(Color.rbTextTertiary)
-                            }
-                        }
-                        Button(action: signInWithGitHub) {
-                            Text("Sign in with GitHub").font(.caption2)
-                        }
-                        .buttonStyle(.bordered)
-                        .help("Authorize RunnerBar via GitHub OAuth and store token in Keychain")
-                    }
-                } else {
-                    // State 4: No token at all.
-                    Button(action: signInWithGitHub) {
-                        Text("Sign in with GitHub").font(.caption2)
-                    }
-                    .buttonStyle(.bordered)
-                    .help("Authorize RunnerBar via GitHub OAuth and store token in Keychain")
-                }
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.vertical, 8)
-        }
-    }
-
-    // MARK: - About
-    private var aboutSection: some View {
-        VStack(alignment: .leading, spacing: 0) {
-            Text("About").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
-                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
-            HStack {
-                Text("Version").font(.system(size: 12)); Spacer()
-                Text("\(appVersion) (\(appBuild))").font(.system(size: 12)).foregroundColor(Color.rbTextSecondary)
-            }
-            .padding(.horizontal, RBSpacing.md).padding(.vertical, 5)
-        }
-    }
-
-    // MARK: - Helpers
-    private func applyLaunchAtLogin(_ enabled: Bool) { LoginItem.setEnabled(enabled) }
-
-    private func signInWithGitHub() {
-        isSigningIn = true
-        // onCompletion is already registered in onAppearAction() — do not re-assign here.
-        OAuthService.shared.signIn()
-    }
-
-    private func signOutOfGitHub() {
-        // OAuthService.signOut() wipes Keychain token and calls onCompletion?(false).
-        // onCompletion re-evaluates both isOAuthAuthenticated and isCLIAuthenticated,
-        // so the UI will flip to CLI state automatically if gh CLI token is still present.
-        OAuthService.shared.signOut()
-    }
-
-    private func performRemoval() {
-        guard let runner = runnerPendingRemoval else { return }
-        runnerPendingRemoval = nil
-        LocalRunnerStore.shared.optimisticallyRemove(runner.runnerName)
-        DispatchQueue.global(qos: .userInitiated).async {
-            let ok = RunnerLifecycleService.shared.remove(runner: runner)
-            DispatchQueue.main.async {
-                if !ok { removeErrorMessage = "Failed to remove \"\(runner.runnerName)\". Check logs." }
                 LocalRunnerStore.shared.refresh()
             }
         }
