+++ Sources/RunnerBar/SettingsViewSections.swift
@@ -1,0 +1,381 @@
+import SwiftUI
+
+// MARK: - SettingsView Sections
+// Section views extracted from SettingsView to keep type_body_length compliant.
+// All state lives in SettingsView; these are pure rendering extensions.
+
+extension SettingsView {
+
+    // MARK: - Sections stack
+
+    var sectionsStack: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            localRunnersSection
+            Divider()
+            remoteScopesSection
+            Divider()
+            notificationsSection
+            Divider()
+            generalSection
+            Divider()
+            accountSection
+            Divider()
+            aboutSection
+        }
+        .padding(.bottom, 16)
+    }
+
+    // MARK: - Header
+
+    var headerBar: some View {
+        HStack {
+            Button(action: onBack, label: {
+                HStack(spacing: 4) {
+                    Image(systemName: "chevron.left")
+                        .font(.system(size: 12, weight: .medium))
+                    Text("Settings").font(.headline)
+                }
+                .foregroundColor(.primary)
+            })
+            .buttonStyle(.plain)
+            Spacer()
+        }
+        .padding(.horizontal, RBSpacing.md).padding(.top, 12).padding(.bottom, 8)
+    }
+
+    // MARK: - Local Runners
+
+    var localRunnersSectionHeader: some View {
+        HStack {
+            Text("Active local runners")
+                .font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+            Spacer()
+            Button(action: { showAddRunnerSheet = true }, label: {
+                Image(systemName: "plus").font(.caption).foregroundColor(Color.rbTextSecondary)
+            })
+            .buttonStyle(.plain).help("Add a new runner").padding(.trailing, 4)
+            if localRunnerStore.isScanning {
+                ProgressView().scaleEffect(0.6).frame(width: 14, height: 14)
+            } else {
+                Button(action: { removeErrorMessage = nil; localRunnerStore.refresh() }, label: {
+                    Image(systemName: "arrow.clockwise").font(.caption).foregroundColor(Color.rbTextSecondary)
+                })
+                .buttonStyle(.plain).help("Refresh local runner list")
+            }
+        }
+        .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+    }
+
+    var localRunnersSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            localRunnersSectionHeader
+            Text("Self-hosted runners installed on this machine, discovered via LaunchAgent plists.")
+                .font(.caption).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
+            if let errMsg = removeErrorMessage {
+                Text(errMsg).font(.caption).foregroundColor(Color.rbDanger)
+                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
+                    .background(Color.rbDanger.opacity(0.07))
+            }
+            if localRunnerStore.runners.isEmpty && !localRunnerStore.isScanning && hasLoadedOnce {
+                Text("No local runners found").font(.caption).foregroundColor(Color.rbTextSecondary)
+                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
+            } else {
+                ForEach(localRunnerStore.runners) { runner in localRunnerRow(runner) }
+            }
+        }
+    }
+
+    func localRunnerRow(_ runner: RunnerModel) -> some View {
+        // swiftlint:disable:next multiple_closures_with_trailing_closure
+        Button(action: { onSelectRunner(runner) }) {
+            localRunnerRowContent(runner)
+                .contentShape(Rectangle())
+        }
+        .buttonStyle(.plain)
+        .padding(.horizontal, RBSpacing.md).padding(.vertical, 5)
+        .background(
+            RoundedRectangle(cornerRadius: RBRadius.small)
+                .fill(Color.rbSurfaceElevated)
+                .overlay(RoundedRectangle(cornerRadius: RBRadius.small)
+                    .strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5))
+                .padding(.horizontal, RBSpacing.xs)
+        )
+    }
+
+    func localRunnerRowContent(_ runner: RunnerModel) -> some View {
+        let hasWarning = runner.lifecycleWarning != nil
+        let displayStatus = runner.displayStatus
+        return HStack(spacing: 6) {
+            Circle().fill(localRunnerDotColor(for: runner)).frame(width: 8, height: 8)
+            VStack(alignment: .leading, spacing: 1) {
+                Text(runner.runnerName).font(.system(size: 13)).lineLimit(1)
+                if let url = runner.gitHubUrl {
+                    Text(url).font(.caption2).foregroundColor(Color.rbTextSecondary).lineLimit(1)
+                }
+            }
+            Spacer()
+            Text(displayStatus)
+                .font(.caption)
+                .foregroundColor(hasWarning ? Color.rbWarning : Color.rbTextSecondary)
+                .lineLimit(1)
+                .fixedSize()
+            if runner.isRunning {
+                Button(action: { performStop(runner: runner) },
+                       label: { Text("Stop").font(.caption2) })
+                .buttonStyle(.bordered).help("Stop runner service")
+            } else {
+                Button(action: { performResume(runner: runner) },
+                       label: { Text("Resume").font(.caption2) })
+                .buttonStyle(.bordered).help("Start runner service")
+            }
+            Image(systemName: "chevron.right")
+                .font(.caption2)
+                .foregroundColor(Color.rbTextTertiary)
+            Button(action: { runnerPendingRemoval = runner },
+                   label: { Image(systemName: "minus.circle").font(.caption2).foregroundColor(Color.rbDanger) })
+            .buttonStyle(.plain).help("Remove runner")
+        }
+    }
+
+    func localRunnerDotColor(for runner: RunnerModel) -> Color {
+        switch runner.statusColor {
+        case .running: return Color.rbSuccess
+        case .busy:    return Color.rbWarning
+        case .idle:    return Color.rbTextTertiary
+        case .offline: return Color.rbDanger
+        }
+    }
+
+    // MARK: - Remote Scopes
+
+    var remoteScopesSectionHeader: some View {
+        HStack {
+            Text("Remote runner scopes")
+                .font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+            Spacer()
+            // swiftlint:disable:next multiple_closures_with_trailing_closure
+            Button(action: { showAddScopeSheet = true }) {
+                Image(systemName: "plus").font(.caption).foregroundColor(Color.rbTextSecondary)
+            }
+            .buttonStyle(.plain).help("Add a remote scope")
+        }
+        .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 2)
+    }
+
+    var remoteScopesSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            remoteScopesSectionHeader
+            Text("GitHub repos or orgs whose runners are fetched via the API.")
+                .font(.caption).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
+            if scopeStore.entries.isEmpty {
+                Text("No remote scopes added")
+                    .font(.caption).foregroundColor(Color.rbTextSecondary)
+                    .padding(.horizontal, RBSpacing.md).padding(.vertical, 4)
+            } else {
+                ForEach(scopeStore.entries) { entry in scopeRow(entry) }
+            }
+        }
+    }
+
+    func scopeRow(_ entry: ScopeEntry) -> some View {
+        let isRepo = entry.scope.contains("/")
+        let displayName = ScopeSettingsStore.displayName(for: entry.scope)
+        // swiftlint:disable:next multiple_closures_with_trailing_closure
+        return Button(action: { onSelectScope(entry) }) {
+            HStack(spacing: 8) {
+                Text(isRepo ? "Repo" : "Org")
+                    .font(.caption2)
+                    .foregroundColor(Color.rbTextSecondary)
+                    .padding(.horizontal, 5)
+                    .padding(.vertical, 2)
+                    .background(Capsule().fill(Color.rbSurfaceElevated))
+                    .overlay(Capsule().strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5))
+                VStack(alignment: .leading, spacing: 1) {
+                    Text(displayName)
+                        .font(.system(size: 12))
+                        .lineLimit(1)
+                        .truncationMode(.middle)
+                    if ScopeSettingsStore.alias(for: entry.scope) != nil {
+                        Text(entry.scope)
+                            .font(.caption2)
+                            .foregroundColor(Color.rbTextTertiary)
+                            .lineLimit(1).truncationMode(.middle)
+                    }
+                }
+                Spacer()
+                Text(entry.isEnabled ? "Active" : "Paused")
+                    .font(.caption2)
+                    .foregroundColor(entry.isEnabled ? Color.rbSuccess : Color.rbTextTertiary)
+                Toggle("", isOn: Binding(
+                    get: { entry.isEnabled },
+                    set: { ScopeStore.shared.setEnabled(entry.id, $0); RunnerStore.shared.start() }
+                ))
+                .toggleStyle(.switch)
+                .tint(Color.rbSuccess)
+                .labelsHidden()
+                .help(entry.isEnabled ? "Pause monitoring" : "Resume monitoring")
+                .scaleEffect(0.8, anchor: .trailing)
+                .buttonStyle(.borderless)
+                Image(systemName: "chevron.right")
+                    .font(.caption2)
+                    .foregroundColor(Color.rbTextTertiary)
+                // swiftlint:disable:next multiple_closures_with_trailing_closure
+                Button(action: {
+                    ScopeSettingsStore.cleanUp(scope: entry.scope)
+                    ScopeStore.shared.remove(id: entry.id)
+                    RunnerStore.shared.start()
+                }) {
+                    Image(systemName: "minus.circle")
+                        .font(.caption2)
+                        .foregroundColor(Color.rbDanger)
+                }
+                .buttonStyle(.borderless)
+                .help("Remove scope")
+            }
+            .contentShape(Rectangle())
+        }
+        .buttonStyle(.plain)
+        .padding(.horizontal, RBSpacing.md)
+        .padding(.vertical, 5)
+        .background(
+            RoundedRectangle(cornerRadius: RBRadius.small)
+                .fill(entry.isEnabled ? Color.rbSurfaceElevated : Color.rbSurface)
+                .overlay(
+                    RoundedRectangle(cornerRadius: RBRadius.small)
+                        .strokeBorder(Color.rbBorderSubtle, lineWidth: 0.5)
+                )
+                .padding(.horizontal, RBSpacing.xs)
+        )
+        .opacity(entry.isEnabled ? 1.0 : 0.5)
+    }
+
+    // MARK: - Notifications
+
+    var notificationsSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            Text("Notifications").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+            HStack {
+                Text("Notify on success").font(.system(size: 12)); Spacer()
+                Toggle("", isOn: $notifications.notifyOnSuccess)
+                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
+            Divider().padding(.leading, RBSpacing.md)
+            HStack {
+                Text("Notify on failure").font(.system(size: 12)); Spacer()
+                Toggle("", isOn: $notifications.notifyOnFailure)
+                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
+        }
+    }
+
+    // MARK: - General
+
+    var generalSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            Text("General").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+            HStack {
+                Text("Launch at login").font(.system(size: 12)); Spacer()
+                Toggle("", isOn: $launchAtLogin)
+                    .toggleStyle(.switch).tint(Color.rbSuccess).labelsHidden()
+                    .onChange(of: launchAtLogin, perform: applyLaunchAtLogin)
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.vertical, 6)
+            Divider().padding(.leading, RBSpacing.md)
+            HStack {
+                Text("Polling interval").font(.system(size: 12)); Spacer()
+                Text("\(settings.pollingInterval)s").font(.system(size: 12)).foregroundColor(Color.rbTextSecondary)
+                    .frame(minWidth: 36, alignment: .trailing)
+                Stepper("", value: $settings.pollingInterval, in: 10...300).labelsHidden()
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.top, 6).padding(.bottom, 2)
+            Text("How often RunnerBar checks GitHub for runner and workflow status. Lower values use more API quota.")
+                .font(.caption).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.bottom, 6)
+        }
+    }
+
+    // MARK: - Account
+
+    var accountSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            Text("Account").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+            HStack(alignment: .center) {
+                Text("GitHub").font(.system(size: 12))
+                Spacer()
+                if isSigningIn {
+                    HStack(spacing: 6) {
+                        ProgressView().scaleEffect(0.7)
+                        Text("Waiting for browser…").font(.caption).foregroundColor(Color.rbTextSecondary)
+                    }
+                } else if isOAuthAuthenticated {
+                    HStack(spacing: 10) {
+                        HStack(spacing: 4) {
+                            Circle().fill(Color.rbSuccess).frame(width: 7, height: 7)
+                            VStack(alignment: .leading, spacing: 1) {
+                                Text("Authenticated")
+                                    .font(.caption)
+                                    .foregroundColor(Color.rbTextSecondary)
+                                Text("via OAuth")
+                                    .font(.caption2)
+                                    .foregroundColor(Color.rbTextTertiary)
+                            }
+                        }
+                        Button(action: signOutOfGitHub) {
+                            Text("Sign out").font(.caption2)
+                        }
+                        .buttonStyle(.bordered)
+                        .tint(Color.rbDanger)
+                        .help("Remove OAuth token from Keychain. gh env token used as fallback if available.")
+                    }
+                } else if isCLIAuthenticated {
+                    HStack(spacing: 10) {
+                        HStack(spacing: 4) {
+                            Circle().fill(Color.rbSuccess).frame(width: 7, height: 7)
+                            VStack(alignment: .leading, spacing: 1) {
+                                Text("Authenticated")
+                                    .font(.caption)
+                                    .foregroundColor(Color.rbTextSecondary)
+                                Text("via gh env token")
+                                    .font(.caption2)
+                                    .foregroundColor(Color.rbTextTertiary)
+                            }
+                        }
+                        Button(action: signInWithGitHub) {
+                            Text("Sign in with GitHub").font(.caption2)
+                        }
+                        .buttonStyle(.bordered)
+                        .help("Authorize RunnerBar via GitHub OAuth and store token in Keychain")
+                    }
+                } else {
+                    Button(action: signInWithGitHub) {
+                        Text("Sign in with GitHub").font(.caption2)
+                    }
+                    .buttonStyle(.bordered)
+                    .help("Authorize RunnerBar via GitHub OAuth and store token in Keychain")
+                }
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.vertical, 8)
+        }
+    }
+
+    // MARK: - About
+
+    var aboutSection: some View {
+        VStack(alignment: .leading, spacing: 0) {
+            Text("About").font(RBFont.sectionHeader).foregroundColor(Color.rbTextSecondary)
+                .padding(.horizontal, RBSpacing.md).padding(.top, 8).padding(.bottom, 4)
+            HStack {
+                Text("Version").font(.system(size: 12)); Spacer()
+                Text("\(appVersion) (\(appBuild))").font(.system(size: 12)).foregroundColor(Color.rbTextSecondary)
+            }
+            .padding(.horizontal, RBSpacing.md).padding(.vertical, 5)
+        }
+    }
+}
